goida

+++ Content.Server/_Shitcode/Heretic/EntitySystems/HereticSystem.cs
@@ -138,7 +138,7 @@ public sealed class HereticSystem : EntitySystem
 
         // add influence layer
         if (TryComp<EyeComponent>(ent, out var eye))
-            _eye.SetVisibilityMask(ent, eye.VisibilityMask | VisibilityFlags.EldritchInfluence);
+            _eye.SetVisibilityMask(ent, eye.VisibilityMask | (int) VisibilityFlags.EldritchInfluence);
 
         foreach (var knowledge in ent.Comp.BaseKnowledge)
             _knowledge.AddKnowledge(ent, ent.Comp, knowledge);
